migrate GetInspections
